+++ donjo-app/models/analisis_master_model.php
@@ -7,7 +7,7 @@ class analisis_master_model extends CI_Model{
 		$sql = "SELECT nama FROM analisis_master";
 		$query = $this->db->query($sql);
 		$data = $query->result_array();
-		
+
 		$i=0;
 		$outp='';
 		while($i<count($data)){
@@ -27,14 +27,14 @@ class analisis_master_model extends CI_Model{
 			return $search_sql;
 			}
 		}
-	function filter_sql(){		
+	function filter_sql(){
 		if(isset($_SESSION['filter'])){
 			$kf = $_SESSION['filter'];
 			$filter_sql= " AND u.subjek_tipe = $kf";
 		return $filter_sql;
 		}
 	}
-	function state_sql(){		
+	function state_sql(){
 		if(isset($_SESSION['state'])){
 			$kf = $_SESSION['state'];
 			$filter_sql= " AND u.lock = $kf";
@@ -43,23 +43,23 @@ class analisis_master_model extends CI_Model{
 	}
 	function paging($p=1,$o=0){
 		$sql = "SELECT COUNT(id) AS id FROM analisis_master u WHERE 1";
-		$sql .= $this->search_sql(); 
-		$sql .= $this->filter_sql(); 
+		$sql .= $this->search_sql();
+		$sql .= $this->filter_sql();
 		$sql .= $this->state_sql();
 		$query = $this->db->query($sql);
 		$row = $query->row_array();
 		$jml_data = $row['id'];
-		
+
 		$this->load->library('paging');
 		$cfg['page'] = $p;
 		$cfg['per_page'] = $_SESSION['per_page'];
 		$cfg['num_rows'] = $jml_data;
 		$this->paging->init($cfg);
-		
+
 		return $this->paging;
 	}
 	function list_data($o=0,$offset=0,$limit=500){
-		
+
 		switch($o){
 			case 1: $order_sql = ' ORDER BY u.nama'; break;
 			case 2: $order_sql = ' ORDER BY u.nama DESC'; break;
@@ -69,9 +69,9 @@ class analisis_master_model extends CI_Model{
 			case 6: $order_sql = ' ORDER BY g.nama DESC'; break;
 			default:$order_sql = ' ORDER BY u.id';
 		}
-		
+
 		$paging_sql = ' LIMIT ' .$offset. ',' .$limit;
-		
+
 		$sql = "SELECT u.*,s.subjek FROM analisis_master u LEFT JOIN analisis_ref_subjek s ON u.subjek_tipe = s.id WHERE 1 ";
 
 		$sql .= $this->search_sql();
@@ -79,10 +79,10 @@ class analisis_master_model extends CI_Model{
 		$sql .= $this->state_sql();
 		$sql .= $order_sql;
 		$sql .= $paging_sql;
-		
+
 		$query = $this->db->query($sql);
 		$data=$query->result_array();
-		
+
 		$i=0;
 		$j=$offset;
 		while($i<count($data)){
@@ -91,7 +91,7 @@ class analisis_master_model extends CI_Model{
 			$data[$i]['lock'] = "<img src='".base_url()."assets/images/icon/unlock.png'>";
 			else
 			$data[$i]['lock'] = "<img src='".base_url()."assets/images/icon/lock.png'>";
-			
+
 			$i++;
 			$j++;
 		}
@@ -100,30 +100,42 @@ class analisis_master_model extends CI_Model{
 	function insert(){
 		$data = $_POST;
 		$outp = $this->db->insert('analisis_master',$data);
-		
+
 		if($outp) $_SESSION['success']=1;
 			else $_SESSION['success']=-1;
 	}
 	function update($id=0){
 		$data = $_POST;
+		// Kolom yang tidak boleh diubah untuk analisis sistem
+		if($this->is_analisis_sistem($id)){
+			unset($data['subjek_tipe']);
+			unset($data['lock']);
+			unset($data['format_impor']);
+		}
 		$this->db->where('id',$id);
 		$outp = $this->db->update('analisis_master',$data);
 		if($outp) $_SESSION['success']=1;
 			else $_SESSION['success']=-1;
 	}
+	function is_analisis_sistem($id){
+		$jenis = $this->db->select('jenis')->where('id',$id)
+			->get('analisis_master')->row()->jenis;
+		return $jenis==1;
+	}
 	function delete($id=''){
-		
+		if($this->is_analisis_sistem($id)) return; // Jangan hapus analisis sistem
+
 		$this->sub_delete($id);
-		
+
 		$sql = "DELETE FROM analisis_master WHERE id=?";
 		$outp = $this->db->query($sql,array($id));
-		
+
 		if($outp) $_SESSION['success']=1;
 			else $_SESSION['success']=-1;
 	}
 	function delete_all(){
 		$id_cb = $_POST['id_cb'];
-		
+
 		if(count($id_cb)){
 			foreach($id_cb as $id){
 				$this->delete($id);
@@ -131,35 +143,35 @@ class analisis_master_model extends CI_Model{
 			$outp = true;
 		}
 		else $outp = false;
-		
+
 		if($outp) $_SESSION['success']=1;
 			else $_SESSION['success']=-1;
 	}
 	function sub_delete($id=''){
-		
+
 		$sql = "DELETE FROM analisis_parameter WHERE id_indikator IN(SELECT id FROM analisis_indikator WHERE id_master = ?)";
 		$this->db->query($sql,$id);
-		
+
 		$sql = "DELETE FROM analisis_respon WHERE id_periode IN(SELECT id FROM analisis_periode WHERE id_master=?)";
 		$this->db->query($sql,$id);
-		
-		
+
+
 		$sql = "DELETE FROM analisis_kategori_indikator WHERE id_master=?";
 		$this->db->query($sql,$id);
-		
+
 		$sql = "DELETE FROM analisis_klasifikasi WHERE id_master=?";
 		$this->db->query($sql,$id);
-		
-		
+
+
 		$sql = "DELETE FROM analisis_respon_hasil WHERE id_master=?";
 		$this->db->query($sql,$id);
-		
+
 		$sql = "DELETE FROM analisis_partisipasi WHERE id_master=?";
 		$this->db->query($sql,$id);
-		
+
 		$sql = "DELETE FROM analisis_periode WHERE id_master=?";
 		$this->db->query($sql,$id);
-		
+
 		$sql = "DELETE FROM analisis_indikator WHERE id_master=?";
 		$this->db->query($sql,$id);
 	}
